Create Students from a file
Write a program that reads characters and strings from the file students.txt and prints them on screen. Follow the format of the example print.

Every student's information is in it's individual row in the file. If the program fails to open the file, it prints “Failed to open file!” on screen. After this the program execution ends.

For example, if the contents of students.txt is:

Z Andy Arkinson

Y Bob Bender

, information is printed as follows:

Andy Arkinson: class Z

Bob Bender: class Y